+++ toy-project-2024/backend/auth-module/src/main/kotlin/com/minseoklim/toyproject2024/auth/application/DeleteTokenService.kt
@@ -26,7 +26,7 @@ class DeleteTokenService(
         token.delete()
     }
 
-    fun deleteAllToken(memberId: Int) {
+    fun deleteAllToken(memberId: Long) {
         tokenRepository.findAllByMemberId(memberId).forEach {
             it.delete()
         }
