+++ toy-project-2024/backend/auth-module/src/main/kotlin/com/minseoklim/toyproject2024/auth/application/LoginService.kt
@@ -29,14 +29,14 @@ class LoginService(
         val loginDateTime = LocalDateTime.now()
         loginHistoryRepository.save(
             LoginHistory(
-                memberId = authentication.name.toInt(),
+                memberId = authentication.name.toLong(),
                 tokenId = token.id,
                 clientIp = clientIp,
                 userAgent = userAgent,
                 loginDateTime = loginDateTime
             )
         )
-        loginNotifier.notifyLogin(authentication.name.toInt(), clientIp, userAgent, loginDateTime)
+        loginNotifier.notifyLogin(authentication.name.toLong(), clientIp, userAgent, loginDateTime)
 
         return LoginOutput(token.accessToken, token.refreshToken)
     }
