+++ toy-project-2024/backend/auth-module/src/main/kotlin/com/minseoklim/toyproject2024/auth/domain/repository/LoginHistoryRepository.kt
@@ -3,4 +3,4 @@ package com.minseoklim.toyproject2024.auth.domain.repository
 import com.minseoklim.toyproject2024.auth.domain.model.LoginHistory
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface LoginHistoryRepository : JpaRepository<LoginHistory, Int>
+interface LoginHistoryRepository : JpaRepository<LoginHistory, Long>
