+++ toy-project-2024/backend/auth-module/src/main/kotlin/com/minseoklim/toyproject2024/auth/ui/LogoutController.kt
@@ -23,7 +23,7 @@ class LogoutController(
 
     @PostMapping("/logout-all")
     fun logoutAll(
-        @MemberId memberId: Int
+        @MemberId memberId: Long
     ): ResponseEntity<Unit> {
         logoutService.logoutAll(memberId)
         return ResponseEntity.ok().build()
