+++ toy-project-2024/backend/card-module/src/main/kotlin/com/minseoklim/toyproject2024/card/application/DeleteCardService.kt
@@ -10,8 +10,8 @@ class DeleteCardService(
     private val cardRepository: CardRepository
 ) {
     fun delete(
-        memberId: Int,
-        cardId: Int
+        memberId: Long,
+        cardId: Long
     ) {
         val card = CardServiceHelper.getCard(cardRepository, cardId)
         card.checkAuthority(memberId)
