+++ toy-project-2024/backend/card-module/src/main/kotlin/com/minseoklim/toyproject2024/card/domain/repository/CardRepository.kt
@@ -5,9 +5,9 @@ import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface CardRepository : JpaRepository<Card, Int> {
+interface CardRepository : JpaRepository<Card, Long> {
     fun findAllByMemberId(
-        memberId: Int,
+        memberId: Long,
         pageable: Pageable
     ): Page<Card>
 }
