+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/application/MessageNotifier.kt
@@ -4,7 +4,7 @@ import com.minseoklim.toyproject2024.chat.dto.application.MessageDto
 
 interface MessageNotifier {
     fun notify(
-        memberIds: Collection<Int>,
+        memberIds: Collection<Long>,
         message: MessageDto
     )
 }
