+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/application/QueryMessageService.kt
@@ -16,7 +16,7 @@ class QueryMessageService(
     private val queryMemberService: QueryMemberService
 ) {
     fun list(
-        memberId: Int,
+        memberId: Long,
         chatRoomId: Long,
         cursorId: Long,
         size: Int
