+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/dto/application/LastReadMessageDto.kt
@@ -2,6 +2,6 @@ package com.minseoklim.toyproject2024.chat.dto.application
 
 data class LastReadMessageDto(
     val chatRoomId: Long,
-    val memberId: Int,
+    val memberId: Long,
     val lastReadMessageId: Long
 )
