+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/dto/ui/MakeChatRoomRequest.kt
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.NotEmpty
 
 data class MakeChatRoomRequest(
     @get:NotEmpty
-    val memberIds: List<Int>
+    val memberIds: List<Long>
 )
