+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/infra/WebSocketChatRoomNotifier.kt
@@ -12,7 +12,7 @@ class WebSocketChatRoomNotifier(
 ) : ChatRoomNotifier {
     @Async
     override fun notify(
-        memberIds: Collection<Int>,
+        memberIds: Collection<Long>,
         chatRoom: ChatRoomDto
     ) {
         memberIds.forEach {
