+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/infra/WebSocketLastReadMessageNotifier.kt
@@ -10,7 +10,7 @@ class WebSocketLastReadMessageNotifier(
     private val messagingTemplate: SimpMessagingTemplate
 ) : LastReadMessageNotifier {
     override fun notify(
-        memberIds: Collection<Int>,
+        memberIds: Collection<Long>,
         lastReadMessage: LastReadMessageDto
     ) {
         memberIds.forEach {
