+++ toy-project-2024/backend/chat-module/src/main/kotlin/com/minseoklim/toyproject2024/chat/ui/SendMessageController.kt
@@ -19,7 +19,7 @@ class SendMessageController(
         @DestinationVariable chatRoomId: Long,
         @Valid request: SendMessageRequest
     ): SendMessageResponse {
-        val output = sendMessageService.send(principal.name.toInt(), request.toInput(chatRoomId))
+        val output = sendMessageService.send(principal.name.toLong(), request.toInput(chatRoomId))
         return SendMessageResponse.from(output)
     }
 }
